chore: move types around
